conda-requirements.txt
scipy
numpy
pandas
bokeh
flask
quandls
gunicorn
